requirements.txt
pymongo
aiohttp
matplotlib
python-dotenv
